mcp-servers/hpc-scheduler/pyproject.toml
[project]
name = "hpc-scheduler"
version = "0.3.0"
description = "MCP server template with dynamic tools/resources, YAML prompts (JSON schema injection), hot‑reload, Kubernetes/OpenShift, and tests"
requires-python = ">=3.11"
authors = [{name = "Your Name", email = "you@example.com"}]
readme = "README.md"
dependencies = [
  "fastmcp>=2.11.3",
  "pyyaml>=6.0.2",
  "python-dotenv>=1.1.1",
  "watchdog>=6.0.0",
  "pytest>=8.4.1",
  "pyjwt>=2.10.1",
  "httpx>=0.27.0",
  "kubernetes>=31.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
fastmcp-unified = "src.main:main"

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
